+++ atlasdb-tests-shared/src/test/java/com/palantir/atlasdb/keyvalue/DirectEncodingMemorySerializableTransactionTest.java
@@ -15,29 +15,10 @@
  */
 package com.palantir.atlasdb.keyvalue;
 
-import com.palantir.atlasdb.AtlasDbConstants;
-import com.palantir.atlasdb.keyvalue.impl.TestResourceManagerV2;
-import com.palantir.atlasdb.transaction.impl.AbstractSerializableTransactionTestV2;
 import com.palantir.atlasdb.transaction.impl.TransactionConstants;
-import com.palantir.atlasdb.transaction.impl.TransactionSchemaVersionEnforcement;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.RegisterExtension;
 
-public class DirectEncodingMemorySerializableTransactionTest extends AbstractSerializableTransactionTestV2 {
-    @RegisterExtension
-    public static final TestResourceManagerV2 TRM = TestResourceManagerV2.inMemory();
-
+public class DirectEncodingMemorySerializableTransactionTest extends AbstractMemorySerializableTransactionTest {
     public DirectEncodingMemorySerializableTransactionTest() {
-        super(TRM, TRM);
-    }
-
-    @BeforeEach
-    public void beforeEach() {
-        keyValueService.truncateTable(AtlasDbConstants.COORDINATION_TABLE);
-        TransactionSchemaVersionEnforcement.ensureTransactionsGoingForwardHaveSchemaVersion(
-                transactionSchemaManager,
-                timestampService,
-                timestampManagementService,
-                TransactionConstants.DIRECT_ENCODING_TRANSACTIONS_SCHEMA_VERSION);
+        super(TransactionConstants.DIRECT_ENCODING_TRANSACTIONS_SCHEMA_VERSION);
     }
 }
